# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BidWrenx API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "422":
          description: Message blocked (circumvention detected)
...
  /messages/threads/{jobId}/{otherUserId}:
    get:
      operationId: getThread
      tags: [messages]
      summary: Get messages in a thread
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
        - name: otherUserId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
...
  /users/me/map-settings:
    patch:
      operationId: updateMyMapSettings
      tags: [auth]
      summary: Update mechanic map presence settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MapSettingsInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, mechanic]
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        warningCount:
          type: integer
        flaggedForReview:
          type: boolean
        suspended:
          type: boolean
        verified:
          type: boolean
        isAdmin:
          type: boolean
        photoUrl:
          type: ["string", "null"]
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        identityVerificationStatus:
          type: string
          enum: [none, pending, approved, rejected]
        insuranceDocumentUploaded:
          type: boolean
        certificationDocumentUploaded:
          type: boolean
        adminVerifiedMechanic:
          type: boolean
        verificationNotes:
          type: ["string", "null"]
        verificationRequestedAt:
          type: ["string", "null"]
        termsAcceptedAt:
          type: ["string", "null"]
        hasPinSet:
          type: boolean
        pinLocked:
          type: boolean
        banned:
          type: boolean
        referralCode:
          type: ["string", "null"]
        specialties:
          type: ["string", "null"]
        legalName:
          type: ["string", "null"]
        displayName:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        showLegalNamePublicly:
          type: boolean
        mapCity:
          type: ["string", "null"]
        mapState:
          type: ["string", "null"]
        mapLat:
          type: ["number", "null"]
        mapLng:
          type: ["number", "null"]
        serviceRadiusMiles:
          type: integer
        isMobileMechanic:
          type: boolean
        shopType:
          type: ["string", "null"]
        willingToTravelMiles:
          type: ["integer", "null"]
        emergencyAvailable:
          type: boolean
        mapVisible:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, role, reviewCount, warningCount, flaggedForReview, suspended, verified, isAdmin, emailVerified, phoneVerified, identityVerificationStatus, insuranceDocumentUploaded, certificationDocumentUploaded, adminVerifiedMechanic, hasPinSet, pinLocked, banned, showLegalNamePublicly, serviceRadiusMiles, isMobileMechanic, emergencyAvailable, mapVisible, createdAt]
...
    MessageInput:
      type: object
      properties:
        jobId:
          type: integer
        recipientId:
          type: integer
        content:
          type: string
      required: [jobId, recipientId, content]
    Message:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        recipientId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
      required: [id, jobId, senderId, senderName, recipientId, content, createdAt]
...
    MapSettingsInput:
      type: object
      properties:
        mapCity:
          type: ["string", "null"]
        mapState:
          type: ["string", "null"]
        serviceRadiusMiles:
          type: integer
        isMobileMechanic:
          type: boolean
        shopType:
          type: string
          enum: [shop, mobile, both]
        willingToTravelMiles:
          type: ["integer", "null"]
        emergencyAvailable:
          type: boolean
        mapVisible:
          type: boolean
...